Remove conflict in tests
Remove conflicts by creating a test suite by Validator and removing
the ValidatorTest test
